pyproject.toml
[tool.poetry]
name = "sample-project"
version = "0.1.0"
description = ""
authors = ["Author Name <author@example.com>"]
# readme = "README.md"
# license = "BSD"
packages = [
  { include = "sample_package" }
]

[tool.poetry.dependencies]
python = "^3.10"
pyopengl = "^3.1.7"
pyopengl-accelerate = "^3.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"